Entity Model for Expenditures

+++ api/test/factories.ts
@@ -1,6 +1,6 @@
 import * as faker from 'faker';
 import { User, UserStatus } from '../models/entity/User';
-import { getConnection, getManager } from 'typeorm';
+import { getConnection } from 'typeorm';
 import { Government } from '../models/entity/Government';
 import { Campaign } from '../models/entity/Campaign';
 import {
@@ -10,6 +10,13 @@ import {
     ContributionType,
     ContributorType
 } from '../models/entity/Contribution';
+import {
+    Expenditure,
+    ExpenditureStatus,
+    ExpenditureSubType,
+    ExpenditureType,
+    PayeeType
+} from '../models/entity/Expenditure';
 
 export async function newActiveUserAsync(): Promise<User> {
     const userRepository = getConnection('default').getRepository('User');
@@ -59,7 +66,7 @@ export async function newCampaignAsync(gov?: Government): Promise<Campaign> {
 export async function newContributionAsync(campaign: Campaign, government: Government): Promise<Contribution> {
     let contribution = new Contribution();
         contribution.address1 = faker.address.streetAddress();
-        contribution.amount = faker.finance.amount(0, 500, 2);
+        contribution.amount = faker.finance.amount(1, 500, 2);
         contribution.campaign = campaign;
         contribution.city = 'Portland';
         contribution.firstName = faker.name.firstName();
@@ -72,18 +79,50 @@ export async function newContributionAsync(campaign: Campaign, government: Gover
         contribution.status = ContributionStatus.DRAFT;
         contribution.zip = '97214';
         contribution.contributorType = ContributorType.INDIVIDUAL;
+        contribution.date = faker.date.past(90);
     const contributionRepository = getConnection('default').getRepository('Contribution');
     contribution = await contributionRepository.save(contribution);
-    console.log('saving contribution', contribution.id);
+    if (process.env.NODE_ENV != 'test') {
+        console.log('saving contribution', contribution.id);
+    }
     return contribution;
+}
+
+export async function newExpenditureAsync(campaign: Campaign, government: Government): Promise<Expenditure> {
+    let expenditure = new Expenditure();
+    expenditure.address1 = faker.address.streetAddress();
+    expenditure.amount = faker.finance.amount(1, 500, 2);
+    expenditure.campaign = campaign;
+    expenditure.city = 'Portland';
+    expenditure.name = faker.name.findName();
+    expenditure.government = government;
+    expenditure.type = ExpenditureType.EXPENDITURE;
+    expenditure.subType = ExpenditureSubType.CASH_EXPENDITURE;
+    expenditure.state = 'OR';
+    expenditure.status = ExpenditureStatus.DRAFT;
+    expenditure.zip = '97214';
+    expenditure.payeeType = PayeeType.INDIVIDUAL;
+    expenditure.description = 'Pays for something';
+    expenditure.date = faker.date.past(90);
+    const expenditureRepository = getConnection('default').getRepository('Expenditure');
+    expenditure = await expenditureRepository.save(expenditure);
+    if (process.env.NODE_ENV != 'test') {
+        console.log('saving expenditure', expenditure.id);
+    }
+    return expenditure;
 }
 
 
 export async function truncateAll() {
+    if (process.env.NODE_ENV === 'production') {
+        throw new Error('Cant run truncate in production');
+    }
     const connection = getConnection('default');
     await connection.query('TRUNCATE "government" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "users" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "campaign" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "permission" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "activity" RESTART IDENTITY CASCADE');
+    await connection.query('TRUNCATE "expenditures" RESTART IDENTITY CASCADE');
+    await connection.query('TRUNCATE "contributions" RESTART IDENTITY CASCADE');
 }
